requirements.txt
Django==3.0.8
Markdown==3.2.2
django-filter==2.3.0
django-taggit==1.3.0
djangorestframework==3.11.0
psycopg2==2.8.5
